[DEBUG] add hook on curl socket close to ensure that the fd is reset to -1
curl socket is stored locally when opened, and used to shutdown the connection. When curl close the
socket, locally stored fd was not reset to -1 and an cancel calls shutdown on a fd that is not owned
anymore.
Add a hook to reset locally stored fd to -1 when curl close the socket

Change-Id: If53332fd1a548a6179e54494aa1fe518bba88a94
Reviewed-on: https://gerrit.parrot.biz/23477
Reviewed-by: Nicolas Brulez <nicolas.brulez@parrot.com>